# localON/docs/openapi.yaml
openapi: 3.1.0
info:
  title: LOCAL ON API
  description: Real-time Seoul citydata based congestion information API
  version: 1.0.0
paths:
  /main:
    get:
      tags:
      - main
      summary: Main Endpoint
      operationId: main_endpoint_main_get
      parameters:
      - name: lat
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          description: User latitude
          title: Lat
        description: User latitude
      - name: lng
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          description: User longitude
          title: Lng
        description: User longitude
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MainOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /areas/{area_id}:
    get:
      tags:
      - areas
      summary: Area Detail Endpoint
      operationId: area_detail_endpoint_areas__area_id__get
      parameters:
      - name: area_id
        in: path
        required: true
        schema:
          type: integer
          exclusiveMinimum: 0
          description: Area ID
          title: Area Id
        description: Area ID
      - name: date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          description: 'Date for hourly timeseries (default: today)'
          title: Date
        description: 'Date for hourly timeseries (default: today)'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /search:
    get:
      tags:
      - search
      summary: Search Endpoint
      operationId: search_endpoint_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          description: Search keyword
          title: Q
        description: Search keyword
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /areas/{area_id}/visit-insights:
    get:
      tags:
      - insights
      summary: Area Visit Insights Endpoint
      operationId: area_visit_insights_endpoint_areas__area_id__visit_insights_get
      parameters:
      - name: area_id
        in: path
        required: true
        schema:
          type: integer
          exclusiveMinimum: 0
          description: Area ID
          title: Area Id
        description: Area ID
      - name: lookback_days
        in: query
        required: false
        schema:
          type: integer
          maximum: 90
          minimum: 7
          description: Historical lookback window
          default: 28
          title: Lookback Days
        description: Historical lookback window
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitInsightOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reviews/reliability:
    post:
      tags:
      - insights
      summary: Review Reliability Endpoint
      operationId: review_reliability_endpoint_reviews_reliability_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewReliabilityIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewReliabilityOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AreaDetailOut:
      properties:
        area_id:
          type: integer
          title: Area Id
        area_cd:
          type: string
          title: Area Cd
        name:
          type: string
          title: Name
        eng_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Eng Name
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        lat:
          anyOf:
          - type: number
          - type: 'null'
          title: Lat
        lng:
          anyOf:
          - type: number
          - type: 'null'
          title: Lng
        congestion:
          $ref: '#/components/schemas/CongestionOut'
        hourly:
          items:
            $ref: '#/components/schemas/HourlyOut'
          type: array
          title: Hourly
        recommendations:
          items:
            $ref: '#/components/schemas/RecommendationOut'
          type: array
          title: Recommendations
        demand:
          $ref: '#/components/schemas/DemandOut'
        commercial:
          $ref: '#/components/schemas/CommercialOut'
        weather:
          $ref: '#/components/schemas/WeatherOut'
      type: object
      required:
      - area_id
      - area_cd
      - name
      - eng_name
      - category
      - lat
      - lng
      - congestion
      - hourly
      - recommendations
      - demand
      - commercial
      - weather
      title: AreaDetailOut
    AreaSummaryOut:
      properties:
        area_id:
          type: integer
          title: Area Id
        area_cd:
          type: string
          title: Area Cd
        name:
          type: string
          title: Name
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        lat:
          anyOf:
          - type: number
          - type: 'null'
          title: Lat
        lng:
          anyOf:
          - type: number
          - type: 'null'
          title: Lng
        congestion_level:
          anyOf:
          - type: string
          - type: 'null'
          title: Congestion Level
        citydata_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Citydata Score
        sdot_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Sdot Score
        population_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Population Min
        population_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Population Max
        weather_temp:
          anyOf:
          - type: number
          - type: 'null'
          title: Weather Temp
        air_idx:
          anyOf:
          - type: string
          - type: 'null'
          title: Air Idx
        distance_m:
          anyOf:
          - type: integer
          - type: 'null'
          title: Distance M
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - area_id
      - area_cd
      - name
      - category
      - lat
      - lng
      - congestion_level
      - citydata_score
      - sdot_score
      - population_min
      - population_max
      - weather_temp
      - air_idx
      - updated_at
      title: AreaSummaryOut
    CommercialOut:
      properties:
        level:
          anyOf:
          - type: string
          - type: 'null'
          title: Level
        payment_cnt:
          anyOf:
          - type: integer
          - type: 'null'
          title: Payment Cnt
        payment_amt_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Payment Amt Min
        payment_amt_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Payment Amt Max
      type: object
      required:
      - level
      - payment_cnt
      - payment_amt_min
      - payment_amt_max
      title: CommercialOut
    CongestionOut:
      properties:
        level:
          anyOf:
          - type: string
          - type: 'null'
          title: Level
        citydata_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Citydata Score
        sdot_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Sdot Score
        msg:
          anyOf:
          - type: string
          - type: 'null'
          title: Msg
        population_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Population Min
        population_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Population Max
        sdot_current:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sdot Current
        sdot_baseline:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sdot Baseline
        is_estimated:
          type: boolean
          title: Is Estimated
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - level
      - citydata_score
      - sdot_score
      - msg
      - population_min
      - population_max
      - sdot_current
      - sdot_baseline
      - is_estimated
      - updated_at
      title: CongestionOut
    DemandOut:
      properties:
        type:
          type: string
          title: Type
        resident_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Resident Rate
        non_resident_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Non Resident Rate
        demographics:
          $ref: '#/components/schemas/DemographicsOut'
      type: object
      required:
      - type
      - resident_rate
      - non_resident_rate
      - demographics
      title: DemandOut
    DemographicsOut:
      properties:
        male_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Male Rate
        female_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Female Rate
        age_0:
          anyOf:
          - type: number
          - type: 'null'
          title: Age 0
        age_10:
          anyOf:
          - type: number
          - type: 'null'
          title: Age 10
        age_20:
          anyOf:
          - type: number
          - type: 'null'
          title: Age 20
        age_30:
          anyOf:
          - type: number
          - type: 'null'
          title: Age 30
        age_40:
          anyOf:
          - type: number
          - type: 'null'
          title: Age 40
        age_50:
          anyOf:
          - type: number
          - type: 'null'
          title: Age 50
        age_60:
          anyOf:
          - type: number
          - type: 'null'
          title: Age 60
        age_70:
          anyOf:
          - type: number
          - type: 'null'
          title: Age 70
      type: object
      required:
      - male_rate
      - female_rate
      - age_0
      - age_10
      - age_20
      - age_30
      - age_40
      - age_50
      - age_60
      - age_70
      title: DemographicsOut
    EventSignalOut:
      properties:
        name:
          type: string
          title: Name
        mention_count:
          type: integer
          title: Mention Count
        period:
          anyOf:
          - type: string
          - type: 'null'
          title: Period
        place:
          anyOf:
          - type: string
          - type: 'null'
          title: Place
      type: object
      required:
      - name
      - mention_count
      - period
      - place
      title: EventSignalOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HotPlaceOut:
      properties:
        rank:
          type: integer
          title: Rank
        area_id:
          type: integer
          title: Area Id
        name:
          type: string
          title: Name
        congestion_level:
          anyOf:
          - type: string
          - type: 'null'
          title: Congestion Level
        citydata_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Citydata Score
        sdot_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Sdot Score
        rank_change:
          anyOf:
          - type: integer
          - type: 'null'
          title: Rank Change
      type: object
      required:
      - rank
      - area_id
      - name
      - congestion_level
      - citydata_score
      - sdot_score
      - rank_change
      title: HotPlaceOut
    HourlyOut:
      properties:
        hour:
          type: integer
          title: Hour
        actual:
          anyOf:
          - type: integer
          - type: 'null'
          title: Actual
        baseline:
          anyOf:
          - type: integer
          - type: 'null'
          title: Baseline
        level:
          anyOf:
          - type: string
          - type: 'null'
          title: Level
        actual_avg:
          anyOf:
          - type: number
          - type: 'null'
          title: Actual Avg
        actual_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Actual Min
        actual_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Actual Max
      type: object
      required:
      - hour
      - actual
      - baseline
      - level
      title: HourlyOut
    MainOut:
      properties:
        areas:
          items:
            $ref: '#/components/schemas/AreaSummaryOut'
          type: array
          title: Areas
        trends:
          $ref: '#/components/schemas/TrendsOut'
      type: object
      required:
      - areas
      - trends
      title: MainOut
    PopulationForecastOut:
      properties:
        at:
          type: string
          format: date-time
          title: At
        expected_population:
          anyOf:
          - type: integer
          - type: 'null'
          title: Expected Population
        confidence:
          type: number
          title: Confidence
        cause_factors:
          items:
            type: string
          type: array
          title: Cause Factors
      type: object
      required:
      - at
      - expected_population
      - confidence
      - cause_factors
      title: PopulationForecastOut
    RecommendationOut:
      properties:
        time_range:
          type: string
          title: Time Range
        expected_level:
          type: string
          title: Expected Level
        avg_population:
          type: integer
          title: Avg Population
      type: object
      required:
      - time_range
      - expected_level
      - avg_population
      title: RecommendationOut
    ReviewItemIn:
      properties:
        text:
          type: string
          maxLength: 3000
          minLength: 1
          title: Text
        rating:
          anyOf:
          - type: number
            maximum: 5.0
            minimum: 0.0
          - type: 'null'
          title: Rating
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
      type: object
      required:
      - text
      title: ReviewItemIn
    ReviewReliabilityIn:
      properties:
        place_name:
          type: string
          maxLength: 200
          minLength: 1
          title: Place Name
        place_id:
          anyOf:
          - type: string
            maxLength: 60
          - type: 'null'
          title: Place Id
        area_id:
          anyOf:
          - type: integer
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Area Id
        source:
          anyOf:
          - type: string
            maxLength: 30
          - type: 'null'
          title: Source
        reviews:
          items:
            $ref: '#/components/schemas/ReviewItemIn'
          type: array
          title: Reviews
      type: object
      required:
      - place_name
      title: ReviewReliabilityIn
    ReviewReliabilityOut:
      properties:
        place_key:
          type: string
          title: Place Key
        place_name:
          type: string
          title: Place Name
        total_reviews:
          type: integer
          title: Total Reviews
        ad_suspect_ratio:
          type: number
          title: Ad Suspect Ratio
        ai_suspect_ratio:
          type: number
          title: Ai Suspect Ratio
        duplicate_ratio:
          type: number
          title: Duplicate Ratio
        trust_score:
          type: number
          title: Trust Score
        grade:
          type: string
          title: Grade
        suspicious_reviews:
          items:
            $ref: '#/components/schemas/SuspiciousReviewOut'
          type: array
          title: Suspicious Reviews
        analyzed_at:
          type: string
          format: date-time
          title: Analyzed At
        model_version:
          type: string
          title: Model Version
      type: object
      required:
      - place_key
      - place_name
      - total_reviews
      - ad_suspect_ratio
      - ai_suspect_ratio
      - duplicate_ratio
      - trust_score
      - grade
      - suspicious_reviews
      - analyzed_at
      - model_version
      title: ReviewReliabilityOut
    RisingOut:
      properties:
        area_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Area Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        change_pct:
          anyOf:
          - type: number
          - type: 'null'
          title: Change Pct
        change_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Change Label
      type: object
      required:
      - area_id
      - name
      - change_pct
      - change_label
      title: RisingOut
    SearchOut:
      properties:
        query:
          type: string
          title: Query
        results:
          items:
            $ref: '#/components/schemas/SearchResultOut'
          type: array
          title: Results
      type: object
      required:
      - query
      - results
      title: SearchOut
    SearchResultOut:
      properties:
        result_type:
          type: string
          enum:
          - localon_area
          - external_place
          title: Result Type
          default: localon_area
        area_id:
          anyOf:
          - type: string
          - type: integer
          title: Area Id
        area_cd:
          anyOf:
          - type: string
          - type: 'null'
          title: Area Cd
        name:
          type: string
          title: Name
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        lat:
          anyOf:
          - type: number
          - type: 'null'
          title: Lat
        lng:
          anyOf:
          - type: number
          - type: 'null'
          title: Lng
        congestion_level:
          anyOf:
          - type: string
          - type: 'null'
          title: Congestion Level
        citydata_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Citydata Score
        sdot_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Sdot Score
        review_trust_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Review Trust Score
        review_trust_grade:
          anyOf:
          - type: string
            enum:
            - high
            - medium
            - low
            - unknown
          - type: 'null'
          title: Review Trust Grade
        review_total_reviews:
          anyOf:
          - type: integer
          - type: 'null'
          title: Review Total Reviews
        review_model_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Review Model Version
        review_analyzed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Review Analyzed At
        review_data_status:
          anyOf:
          - type: string
            enum:
            - analyzed
            - insufficient_data
          - type: 'null'
          title: Review Data Status
      type: object
      required:
      - area_id
      - name
      - category
      - lat
      - lng
      title: SearchResultOut
    SuspiciousReviewOut:
      properties:
        reason:
          type: string
          title: Reason
        preview:
          type: string
          title: Preview
      type: object
      required:
      - reason
      - preview
      title: SuspiciousReviewOut
    TrendsOut:
      properties:
        hot_places:
          items:
            $ref: '#/components/schemas/HotPlaceOut'
          type: array
          title: Hot Places
        rising:
          items:
            $ref: '#/components/schemas/RisingOut'
          type: array
          title: Rising
        popular_searches:
          items:
            type: string
          type: array
          title: Popular Searches
      type: object
      required:
      - hot_places
      - rising
      - popular_searches
      title: TrendsOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VisitInsightOut:
      properties:
        area_id:
          type: integer
          title: Area Id
        area_name:
          type: string
          title: Area Name
        lookback_days:
          type: integer
          title: Lookback Days
        generated_at:
          type: string
          format: date-time
          title: Generated At
        hottest_slots:
          items:
            $ref: '#/components/schemas/VisitTimeSlotOut'
          type: array
          title: Hottest Slots
        recommended_slots:
          items:
            $ref: '#/components/schemas/VisitTimeSlotOut'
          type: array
          title: Recommended Slots
        next_hours_forecast:
          items:
            $ref: '#/components/schemas/PopulationForecastOut'
          type: array
          title: Next Hours Forecast
        event_signals:
          items:
            $ref: '#/components/schemas/EventSignalOut'
          type: array
          title: Event Signals
      type: object
      required:
      - area_id
      - area_name
      - lookback_days
      - generated_at
      - hottest_slots
      - recommended_slots
      - next_hours_forecast
      - event_signals
      title: VisitInsightOut
    VisitTimeSlotOut:
      properties:
        day_of_week:
          type: integer
          title: Day Of Week
        day_label:
          type: string
          title: Day Label
        hour:
          type: integer
          title: Hour
        time_range:
          type: string
          title: Time Range
        avg_population:
          anyOf:
          - type: integer
          - type: 'null'
          title: Avg Population
        congestion_ratio:
          anyOf:
          - type: number
          - type: 'null'
          title: Congestion Ratio
        hot_score:
          type: number
          title: Hot Score
        visit_score:
          type: number
          title: Visit Score
        confidence:
          type: number
          title: Confidence
        primary_cause:
          anyOf:
          - type: string
          - type: 'null'
          title: Primary Cause
      type: object
      required:
      - day_of_week
      - day_label
      - hour
      - time_range
      - avg_population
      - congestion_ratio
      - hot_score
      - visit_score
      - confidence
      - primary_cause
      title: VisitTimeSlotOut
    WeatherOut:
      properties:
        temp:
          anyOf:
          - type: number
          - type: 'null'
          title: Temp
        sensible_temp:
          anyOf:
          - type: number
          - type: 'null'
          title: Sensible Temp
        humidity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Humidity
        wind_spd:
          anyOf:
          - type: number
          - type: 'null'
          title: Wind Spd
        pm25:
          anyOf:
          - type: number
          - type: 'null'
          title: Pm25
        pm25_index:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm25 Index
        pm10:
          anyOf:
          - type: number
          - type: 'null'
          title: Pm10
        pm10_index:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm10 Index
      type: object
      required:
      - temp
      - sensible_temp
      - humidity
      - wind_spd
      - pm25
      - pm25_index
      - pm10
      - pm10_index
      title: WeatherOut